Implement para algorithm
The Algebra of a cata exposes the functor full of computed values. This
is what you want when you are folding up from primitive leafs. But if
you need to perform computations over internal nodes, then you need a
para. A simple example is an algorithm for finding the maximum number of
non-leaf children of a single node. At each node you need to count the
number of non-leaf children and the maxes from each child, then return
the highest.

A cata operates on folded children -- it cannot peak into what they were
before the computation completed. A para has access to both the original
structure AND the folded structure.

For this simple example, explicit recursion is actually simpler, since
it deals with the simpler (Tree a b c) rather than (TreeF a b c r)
structure (and the associated In's and out's).