Add list para and ana, merge rs with alg
It may be useful to define algebras as independent functions, sometimes,
but more often it makes sense to bind them in a single function with
their recursion scheme.

This will be the most common recursive pattern:

```
foo :: Term (F a b) -> X
foo = cata go where
    go (Node ...) = ...
```

This way we don't have to write the cata/ana or whatever at every use
case, instead it is hidden in the implementation, as it should be.

I could write vim snippets to expand this pattern for each of the
schemes.

Type aliases could also be used to hide the Term's:

```
TreeF' a b c r = Leaf a | Node b [(c, r)]

type TreeF = Term (TreeF' a b c)
```

This is cute, but it also hides reality without adding real abstaction.