Write Nat with factorial example
The factorial example requires a paramorphism. A catamorphisms knows the
result of the child factorial, but doesn't know what to multipley by.
However, the paramorphism is inefficient since (among other reasons) it
must reconvert Nat to Integer every step. I could reimplement it with a
histomorphism that caches the computation.