Tested ukp3 with larger numbers of components
Hmm, the knapsack problem seems to be related to the factorization
problem. Or I guess it is the factorization problem for sums where there
are constraints on which values may be used.

Is there sum combination of [x1, x2, ..., x_n] that sum to N? That is
the decision knapsack problem. The optimization problem is, "What is the
smallest combination of [x1j, x2, ..., x_n] that sum to N?"