Wrote general unbounded knapsack solution
Each recursive step needs to choose among all children which is *best*.
For the coin change problem, the best choice is the one where change
exactly hits the goal and has the minimum number of coins.

But other problems, like filling a knapsack, might instead want to
maximize the number of items in the sack, rather than minimize the
remainder.

Or the goal could be to minimize remainder but resolve ties with the
maximum number of elements.

The general function has a criterion function as the first argument. All
the cases above, and more, can be implemented by changing the criterion
function. Here is the type signature for all the criteria:

```
criterion :: (Num weight, Ord weight, Ord cost)
          => (weight, [weight]) -> cost
```

Here are a few options for criteria:

Minimize remainder first, then minimize number of items. This is the
criteria for the exact change problem:

```
criterion1 (r, xs) = (r, length xs)
```

Minimize remainder first, then maximize number of items.

```
criterion2 (r, xs) = (r, (-1) * length xs)
```

Or burglar style, throw some big things in a bag and run

```
criterion2 (r, xs) = (-1) * mean xs + r
```