(n log(n)) and (n * k) solutions to KLV
The kth largest value (KLV) problem requires searching through an
unsorted list and finding the kth largest value.

The easiest solution is to sort the list and then find the kth value.
This runs in O(n log(n)) time due to the sorting step.

Alternatively, you can maintain a list of length k to which elements are
merged upwards if the are greater than the first element in the list
(k). This has a worst case runtime of O(n * k). The worst case is when
the input list is ordered and element in xs must be pushed all the way
to the end of the k-length list. If xs is random and long, performance
is better since increasingly few elements need to be pushed far into the
ks list.

On randomly ordered data, the O(n * k) algorithm seems nearly
independent of k (at least my sloppy benchmarks aren't conclusive). This
makes sense because the list of the highest k elements quickly saturates
and increasingly few elements need to be added.

A better solution to this problem would use a tree instead of a list.
then worst case performance would be O(n (log_b k)), where b is the base
of the tree. For a finger tree, b is large, meaning runtime could be
close to O(n).

So I need to learn how finger trees work. But before doing that, I
should write up everything I learned from the knapsack problem and
finish the histomorphism implementation.